Replaced old error indicator function
Replaced the old code for putting the MCU to deep sleep with orange led. Now it's not dependent of the normal function loop which could cause it to hang again when an AD conversion takes to long time for example. It's now an independent check and everything is done before entering the main loop.